Added a double specialization to AbsPrim and move it
