+++ src/som/primitives/arithmetic/AbsPrim.java
@@ -1,0 +1,33 @@
+package som.primitives.arithmetic;
+
+import com.oracle.truffle.api.dsl.GenerateNodeFactory;
+import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.instrumentation.Tag;
+
+import bd.primitives.Primitive;
+import som.interpreter.nodes.nary.UnaryBasicOperation;
+import tools.dym.Tags.OpArithmetic;
+
+
+@GenerateNodeFactory
+@Primitive(selector = "abs", receiverType = {Long.class, Double.class})
+public abstract class AbsPrim extends UnaryBasicOperation {
+  @Override
+  protected boolean hasTagIgnoringEagerness(final Class<? extends Tag> tag) {
+    if (tag == OpArithmetic.class) {
+      return true;
+    } else {
+      return super.hasTagIgnoringEagerness(tag);
+    }
+  }
+
+  @Specialization
+  public final long doLong(final long receiver) {
+    return Math.abs(receiver);
+  }
+
+  @Specialization
+  public final double doDouble(final double receiver) {
+    return Math.abs(receiver);
+  }
+}
